refactor: extract ui components and views
