+++ client/src/components/game/GameBoard.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { useState, useCallback } from "react";
-import { CellState, XorO } from "../types";
+import { CellState, XorO } from "../../types";
+import { GameCell } from "./GameCell";
 
 type GameBoardProps = {
   board: CellState[];
@@ -30,10 +31,12 @@ export function GameBoard({ board, setBoard }: GameBoardProps) {
 
   return (
     <div className="flex flex-col mt-10 items-center gap-10">
-      <h1 className="font-bold text-4xl tracking-tight text-white">Tic Tac Toe</h1>
+      <h1 className="font-bold text-4xl tracking-tight text-white">
+        Tic Tac Toe
+      </h1>
       <div className="grid grid-cols-3 gap-2 p-6 bg-background-secondary rounded-lg shadow-xl">
         {board.map((cell, idx) => (
-          <GameCellMemo
+          <GameCell
             key={idx}
             cellState={cell}
             cellIndex={idx}
@@ -41,29 +44,6 @@ export function GameBoard({ board, setBoard }: GameBoardProps) {
           />
         ))}
       </div>
-    </div>
-  );
-}
-
-type GameCellProps = {
-  cellState: CellState;
-  cellIndex: number;
-  onCellClick: (cellIdx: number) => void;
-};
-
-const GameCellMemo = React.memo(GameCell);
-function GameCell({
-  cellState,
-  cellIndex,
-  onCellClick,
-}: GameCellProps): React.JSX.Element {
-  return (
-    <div
-      className="border-2 border-border-primary bg-background-primary w-20 h-20 cursor-pointer items-center justify-center text-3xl font-bold flex rounded-md transition-all duration-200 hover:bg-background-secondary hover:border-accent-sage hover:scale-105 active:scale-95 text-border-light"
-      data-testid={`cell-${cellIndex}`}
-      onClick={() => onCellClick(cellIndex)}
-    >
-      {cellState}
     </div>
   );
 }
