+++ client/src/main.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useMemo } from "react";
 import { CellState } from "./types";
-import { OutcomeModal } from "./components/OutcomeModal";
 import { getIfWinner } from "./helpers/winning-conditions-helpers";
-import { GameBoard } from "./components/GameBoard";
+import { GameBoard } from "./components/game/GameBoard";
+import { OutcomeView } from "./components/OutcomeView";
 
 export const BOARD_LENGTH = 3;
 const EMPTY_BOARD = Array(BOARD_LENGTH * BOARD_LENGTH).fill(undefined);
@@ -21,7 +21,7 @@ export const Main = () => {
   return (
     <>
       {<GameBoard key={resetKey} board={board} setBoard={setBoard} />}
-      {winner && <OutcomeModal outcome={winner} resetBoard={resetBoard} />}
+      {winner && <OutcomeView outcome={winner} resetBoard={resetBoard} />}
     </>
   );
 };
